+++ lua/fzfx/config.lua
@@ -2339,228 +2339,7 @@ local Defaults = {
   files = require("fzfx.cfg.files"),
 
   -- the 'Live Grep' commands
-  --- @type fzfx.GroupConfig
-  live_grep = {
-    commands = {
-      -- normal
-      {
-        name = "FzfxLiveGrep",
-        feed = CommandFeedEnum.ARGS,
-        opts = {
-          bang = true,
-          nargs = "*",
-          desc = "Live grep",
-        },
-        default_provider = "restricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepU",
-        feed = CommandFeedEnum.ARGS,
-        opts = {
-          bang = true,
-          nargs = "*",
-          desc = "Live grep unrestricted",
-        },
-        default_provider = "unrestricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepB",
-        feed = CommandFeedEnum.ARGS,
-        opts = {
-          bang = true,
-          nargs = "*",
-          desc = "Live grep on current buffer",
-        },
-        default_provider = "buffer_mode",
-      },
-      -- visual
-      {
-        name = "FzfxLiveGrepV",
-        feed = CommandFeedEnum.VISUAL,
-        opts = {
-          bang = true,
-          range = true,
-          desc = "Live grep by visual select",
-        },
-        default_provider = "restricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepUV",
-        feed = CommandFeedEnum.VISUAL,
-        opts = {
-          bang = true,
-          range = true,
-          desc = "Live grep unrestricted by visual select",
-        },
-        default_provider = "unrestricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepBV",
-        feed = CommandFeedEnum.VISUAL,
-        opts = {
-          bang = true,
-          range = true,
-          desc = "Live grep on current buffer by visual select",
-        },
-        default_provider = "buffer_mode",
-      },
-      -- cword
-      {
-        name = "FzfxLiveGrepW",
-        feed = CommandFeedEnum.CWORD,
-        opts = {
-          bang = true,
-          desc = "Live grep by cursor word",
-        },
-        default_provider = "restricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepUW",
-        feed = CommandFeedEnum.CWORD,
-        opts = {
-          bang = true,
-          desc = "Live grep unrestricted by cursor word",
-        },
-        default_provider = "unrestricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepBW",
-        feed = CommandFeedEnum.CWORD,
-        opts = {
-          bang = true,
-          desc = "Live grep on current buffer by cursor word",
-        },
-        default_provider = "buffer_mode",
-      },
-      -- put
-      {
-        name = "FzfxLiveGrepP",
-        feed = CommandFeedEnum.PUT,
-        opts = {
-          bang = true,
-          desc = "Live grep by yank text",
-        },
-        default_provider = "restricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepUP",
-        feed = CommandFeedEnum.PUT,
-        opts = {
-          bang = true,
-          desc = "Live grep unrestricted by yank text",
-        },
-        default_provider = "unrestricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepBP",
-        feed = CommandFeedEnum.PUT,
-        opts = {
-          bang = true,
-          desc = "Live grep on current buffer by yank text",
-        },
-        default_provider = "buffer_mode",
-      },
-      -- resume
-      {
-        name = "FzfxLiveGrepR",
-        feed = CommandFeedEnum.RESUME,
-        opts = {
-          bang = true,
-          desc = "Live grep by resume last",
-        },
-        default_provider = "restricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepUR",
-        feed = CommandFeedEnum.RESUME,
-        opts = {
-          bang = true,
-          desc = "Live grep unrestricted by resume last",
-        },
-        default_provider = "unrestricted_mode",
-      },
-      {
-        name = "FzfxLiveGrepBR",
-        feed = CommandFeedEnum.RESUME,
-        opts = {
-          bang = true,
-          desc = "Live grep on current buffer by resume last",
-        },
-        default_provider = "buffer_mode",
-      },
-    },
-    providers = {
-      restricted_mode = {
-        key = "ctrl-r",
-        provider = providers_helper.provide_live_grep_restricted_mode,
-        provider_type = ProviderTypeEnum.COMMAND_LIST,
-        line_opts = {
-          prepend_icon_by_ft = true,
-          prepend_icon_path_delimiter = ":",
-          prepend_icon_path_position = 1,
-        },
-      },
-      unrestricted_mode = {
-        key = "ctrl-u",
-        provider = providers_helper.provide_live_grep_unrestricted_mode,
-        provider_type = ProviderTypeEnum.COMMAND_LIST,
-        line_opts = {
-          prepend_icon_by_ft = true,
-          prepend_icon_path_delimiter = ":",
-          prepend_icon_path_position = 1,
-        },
-      },
-      buffer_mode = {
-        key = "ctrl-o",
-        provider = providers_helper.provide_live_grep_buffer_mode,
-        provider_type = ProviderTypeEnum.COMMAND_LIST,
-        line_opts = {
-          prepend_icon_by_ft = true,
-          prepend_icon_path_delimiter = ":",
-          prepend_icon_path_position = 1,
-        },
-      },
-    },
-    previewers = {
-      restricted_mode = {
-        previewer = previewers_helper.preview_live_grep_restricted_mode,
-        previewer_type = PreviewerTypeEnum.COMMAND_LIST,
-        previewer_label = consts.HAS_RG and labels_helper.label_rg
-          or labels_helper.label_grep,
-      },
-      unrestricted_mode = {
-        previewer = previewers_helper.preview_live_grep_unrestricted_mode,
-        previewer_type = PreviewerTypeEnum.COMMAND_LIST,
-        previewer_label = consts.HAS_RG and labels_helper.label_rg
-          or labels_helper.label_grep,
-      },
-      buffer_mode = {
-        previewer = previewers_helper.preview_live_grep_buffer_mode,
-        previewer_type = PreviewerTypeEnum.COMMAND_LIST,
-        previewer_label = consts.HAS_RG and labels_helper.label_rg
-          or labels_helper.label_grep,
-      },
-    },
-    actions = {
-      ["esc"] = actions_helper.nop,
-      ["enter"] = consts.HAS_RG and actions_helper.edit_rg
-        or actions_helper.edit_grep,
-      ["double-click"] = consts.HAS_RG and actions_helper.edit_rg
-        or actions_helper.edit_grep,
-      ["ctrl-q"] = consts.HAS_RG and actions_helper.setqflist_rg
-        or actions_helper.setqflist_grep,
-    },
-    fzf_opts = {
-      default_fzf_options.multi,
-      "--disabled",
-      { "--prompt", "Live Grep > " },
-      { "--delimiter", ":" },
-      { "--preview-window", "+{2}-/2" },
-    },
-    other_opts = {
-      reload_on_change = true,
-    },
-  },
+  live_grep = require("fzfx.cfg.live_grep"),
 
   -- the 'Buffers' commands
   --- @type fzfx.GroupConfig
